Fixed crash during save after DRC on Java 1.6
git-svn-id: https://svn.java.net/svn/electric~svn/trunk@5859 b64ed6c1-75a8-e629-b1da-d0dae74a743a

+++ srcj/com/sun/electric/tool/io/output/ELIB.java
@@ -1173,7 +1173,7 @@ public class ELIB extends Output
             return;
         }
         if (obj instanceof Long) {
-            writeBigInteger(((Integer)obj).intValue());
+            writeBigInteger(((Long)obj).intValue());
             return;
         }
         if (obj instanceof Integer) {
